refactor(nextcloud): Move version support to helpers
