+++ packages/nextcloud/lib/nextcloud.dart
@@ -25,8 +25,8 @@ export 'src/app_type.dart';
 export 'src/client.dart';
 export 'src/helpers/core.dart';
 export 'src/helpers/news.dart';
+export 'src/helpers/notes.dart';
 export 'src/helpers/notifications.dart';
-export 'src/version_supported.dart';
 export 'src/webdav/client.dart';
 export 'src/webdav/file.dart';
 export 'src/webdav/props.dart';
